칭호완료데이터 추가
0안 붙었을때 정렬 안되는 문제 해결
