+++ app/src/main/java/com/abbsolute/ma_livu/Home/ToDoList/ToDoAdapter.java
@@ -16,9 +16,19 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.abbsolute.ma_livu.Firebase.FirebaseID;
 import com.abbsolute.ma_livu.R;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.SetOptions;
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 public class ToDoAdapter extends RecyclerView.Adapter<ToDoAdapter.ViewHolder> {
     ArrayList<ToDoInfo> arrayList = new ArrayList<>();
@@ -28,6 +38,8 @@ public class ToDoAdapter extends RecyclerView.Adapter<ToDoAdapter.ViewHolder> {
     // CheckBox checkBox;
     boolean check = false;
     private OnToDoTextClick onCallBack;
+    private FirebaseFirestore firestore = FirebaseFirestore.getInstance();
+    private FirebaseAuth firebaseAuth = FirebaseAuth.getInstance();
     public class ViewHolder extends RecyclerView.ViewHolder {
         protected TextView ContentsDetail;
         protected TextView Contents;
@@ -113,6 +125,132 @@ public class ToDoAdapter extends RecyclerView.Adapter<ToDoAdapter.ViewHolder> {
         SharedPreferences pf = context.getSharedPreferences("pref", Activity.MODE_PRIVATE);
         Boolean chk1 = pf.getBoolean("chk" + position, false);
         holder.checkToDo.setChecked(chk1);
+        if(chk1==true){//완료했다는 표시
+            String todo=arrayList.get(position).getContent();
+            String detailToDo=arrayList.get(position).getDetailContent();
+            if(todo.equals("청소")){
+                final String[] newCleaningNum = {""};
+                final boolean[] flag = {false};
+                DocumentReference documentReference=firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid());
+                documentReference.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                        if(task.isSuccessful()){
+                            DocumentSnapshot snapshot=task.getResult();
+                            String tmp= (String) snapshot.getData().get("cleaning");
+                            int tmpNum=Integer.parseInt(tmp);
+                            if(tmpNum>=1){
+                                String cleaningNum= (String) snapshot.getData().get("cleaning");
+                                Toast.makeText(context, cleaningNum, Toast.LENGTH_SHORT).show();
+                                int tmpCleaningNum=0;
+                                try {
+                                    tmpCleaningNum=Integer.parseInt(cleaningNum)+1;
+                                }catch (NumberFormatException e){
+                                    e.printStackTrace();
+                                }
+                                String newCleaningNum=Integer.toString(tmpCleaningNum);
+                                if(!newCleaningNum.equals("1")){
+                                    Map<String, Object> data = new HashMap<>();
+                                    //data.put(FirebaseID.documentID,firebaseAuth.getCurrentUser().getUid());
+                                    data.put(FirebaseID.cleaning, newCleaningNum);
+                                    firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid()).set(data, SetOptions.merge());
+                                    Toast.makeText(context, "청소 칭호 카운트 업!!", Toast.LENGTH_SHORT).show();
+                                }
+                            }else{
+                                newCleaningNum[0]="1";
+                                Map<String, Object> data = new HashMap<>();
+                                // data.put(FirebaseID.documentID,firebaseAuth.getCurrentUser().getUid());
+                                data.put(FirebaseID.cleaning, newCleaningNum[0]);
+                                firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid()).set(data, SetOptions.merge());
+                                Toast.makeText(context, "청소 칭호 카운트 업!!", Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    }
+                });
+
+            }//청소하기 완료!!!
+            else if(todo.equals("빨래")){
+                final String[] newLaundryNum = {""};
+                final boolean[] flag = {false};
+                DocumentReference documentReference=firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid());
+                documentReference.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                        if(task.isSuccessful()){
+                            DocumentSnapshot snapshot=task.getResult();
+                            String tmp= (String) snapshot.getData().get("laundry");//횟수가 1이상인지 확인하는 코드
+                            int tmpNum=Integer.parseInt(tmp);
+                            if(tmpNum>=1){
+                                String laundryNum= (String) snapshot.getData().get("laundry");
+                                Toast.makeText(context, laundryNum, Toast.LENGTH_SHORT).show();
+                                int tmpLaundryNum=0;
+                                try{
+                                    tmpLaundryNum=Integer.parseInt(laundryNum)+1;
+                                }catch (NumberFormatException e){
+                                    e.printStackTrace();
+                                }
+                                String newLaundryNum=Integer.toString(tmpLaundryNum);
+                                if(!newLaundryNum.equals("1")){
+                                    Map<String, Object> data = new HashMap<>();
+                                    //data.put(FirebaseID.documentID,firebaseAuth.getCurrentUser().getUid());
+                                    data.put(FirebaseID.laundry, newLaundryNum);
+                                    firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid()).set(data, SetOptions.merge());
+                                    Toast.makeText(context, "빨래하기 칭호 카운트 업!!", Toast.LENGTH_SHORT).show();
+                                }
+                            }else{
+                                newLaundryNum[0]="1";
+                                Map<String, Object> data = new HashMap<>();
+                                // data.put(FirebaseID.documentID,firebaseAuth.getCurrentUser().getUid());
+                                data.put(FirebaseID.laundry, newLaundryNum[0]);
+                                firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid()).set(data, SetOptions.merge());
+                                Toast.makeText(context, "빨래하기 칭호 카운트 업!!", Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    }
+                });
+
+            }//빨래하기 완료!!
+            else if(todo.equals("쓰레기")){
+                final String[] newtrashNum = {""};
+                DocumentReference documentReference=firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid());
+                documentReference.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                        if(task.isSuccessful()){
+                            DocumentSnapshot snapshot=task.getResult();
+                            String tmp= (String) snapshot.getData().get("trash");//횟수가 1이상인지 확인하는 코드
+                            int tmpNum=Integer.parseInt(tmp);
+                            if(tmpNum>=1){
+                                String trashNum= (String) snapshot.getData().get("trash");
+                                Toast.makeText(context, trashNum, Toast.LENGTH_SHORT).show();
+                                int tmpTrashNum=0;
+                                try{
+                                    tmpTrashNum=Integer.parseInt(trashNum)+1;
+                                }catch (NumberFormatException e){
+                                    e.printStackTrace();
+                                }
+                                String newTrashNum=Integer.toString(tmpTrashNum);
+                                if(!newTrashNum.equals("1")){
+                                    Map<String, Object> data = new HashMap<>();
+                                    // data.put(FirebaseID.documentID,firebaseAuth.getCurrentUser().getUid());
+                                    data.put(FirebaseID.ToDoTrash, newTrashNum);
+                                    firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid()).set(data, SetOptions.merge());
+                                    Toast.makeText(context, "쓰레기 칭호 카운트 업!!", Toast.LENGTH_SHORT).show();
+                                }
+                            }else{
+                                newtrashNum[0]="1";
+                                Map<String, Object> data = new HashMap<>();
+                                //data.put(FirebaseID.documentID,firebaseAuth.getCurrentUser().getUid());
+                                data.put(FirebaseID.ToDoTrash, newtrashNum[0]);
+                                firestore.collection(FirebaseID.ToDoLists).document(firebaseAuth.getCurrentUser().getUid()).set(data, SetOptions.merge());
+                                Toast.makeText(context, "쓰레기 칭호 카운트 업!!", Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    }
+                });
+                
+            }//쓰레기 완료!!!
+        }
         holder.checkToDo.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
